Fix lateral speed/accel data types

+++ src/vt/tr_types.h
@@ -455,8 +455,8 @@ typedef struct
     int32_t   speed;
     int32_t   accel;
 
-    float   speed_lateral;      // new in TR4 -->
-    float   accel_lateral;      // lateral speed and acceleration.
+    int32_t   speed_lateral;      // new in TR4 -->
+    int32_t   accel_lateral;      // lateral speed and acceleration.
 
     uint16_t frame_start;           // first frame in this animation
     uint16_t frame_end;             // last frame in this animation (numframes = (End - Start) + 1)
